Split engine tests by responsibility
